pyproject.toml
[project]
name = "ir-bot"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.9"
dependencies = [
    "bs4>=0.0.2",
    "discord-py>=2.5.2",
    "google-auth>=2.40.3",
    "gspread>=6.0.2",
    "gspread-asyncio==2.0.0",
    "html5lib>=1.1",
    "ipykernel>=6.30.0",
    "lxml>=6.0.0",
    "matplotlib>=3.9.4",
    "oauth2client>=4.1.3",
    "pandas>=2.3.1",
    "python-dotenv>=1.1.1",
    "requests>=2.32.4",
    "scipy>=1.13.1",
    "tabulate>=0.9.0",
    "tqdm>=4.67.1",
]
